Re-enable controller sounds and improve, fix mesh intensity transition alpha bug

+++ lua/photon-v2/cl_render_light_mesh.lua
@@ -54,7 +54,9 @@ function Photon2.RenderLightMesh.Render( depth, sky )
 			-- light.DrawMaterial--[[@as IMaterial]]:SetVector( "$color", Vector( 0, 0, 1 ) )
 			light.DrawMaterial--[[@as IMaterial]]:SetVector( "$color", light.DrawColor:GetVector() )
 			
-			light.DrawMaterial--[[@as IMaterial]]:SetInt( "$alpha", light.Intensity )
+			-- light.DrawMaterial--[[@as IMaterial]]:SetFloat( "$alpha", 0 )
+			light.DrawMaterial--[[@as IMaterial]]:SetFloat( "$alpha", light.Intensity )
+			-- light.DrawMaterial--[[@as IMaterial]]:SetInt( "$alpha", light.Intensity )
 			-- local cMatrix = Matrix({{512, 512, 512, 512}, {512, 512, 512, 512}, {512,512,512,512}, {512,512,512,512}})
 			-- cMatrix:Translate(Vector(1024, 1024, 0))
 			-- l
